[BNK-179] Rename: 상품 경로 수정
