+++ src/main/java/com/banklab/product/batch/tasklet/creditloan/DeleteCreditLoanTasklet.java
@@ -1,9 +1,7 @@
-package com.banklab.product.batch.tasklet;
+package com.banklab.product.batch.tasklet.creditloan;
 
-import com.banklab.product.mapper.DepositOptionMapper;
-import com.banklab.product.mapper.DepositProductMapper;
-import com.banklab.product.mapper.SavingsOptionMapper;
-import com.banklab.product.mapper.SavingsProductMapper;
+import com.banklab.product.mapper.CreditLoanOptionMapper;
+import com.banklab.product.mapper.CreditLoanProductMapper;
 import org.springframework.batch.core.StepContribution;
 import org.springframework.batch.core.scope.context.ChunkContext;
 import org.springframework.batch.core.step.tasklet.Tasklet;
@@ -12,18 +10,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
-public class DeleteSavingsTasklet implements Tasklet {
+public class DeleteCreditLoanTasklet implements Tasklet {
 
     @Autowired
-    private SavingsProductMapper savingsProductMapper;
+    private CreditLoanProductMapper creditLoanProductMapper;
+    
     @Autowired
-    private SavingsOptionMapper savingsOptionMapper;
+    private CreditLoanOptionMapper creditLoanOptionMapper;
 
     @Override
     public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
-        savingsProductMapper.deleteAllSavingsProducts();
-        savingsOptionMapper.deleteAllSavingsOptions();
+        creditLoanProductMapper.deleteAllCreditLoanProducts();
+        creditLoanOptionMapper.deleteAllCreditLoanOptions();
         return RepeatStatus.FINISHED;
     }
 }
-
