+++ src/main/java/com/banklab/product/batch/tasklet/deposit/DeleteDepositTasklet.java
@@ -1,7 +1,7 @@
-package com.banklab.product.batch.tasklet;
+package com.banklab.product.batch.tasklet.deposit;
 
-import com.banklab.product.mapper.CreditLoanOptionMapper;
-import com.banklab.product.mapper.CreditLoanProductMapper;
+import com.banklab.product.mapper.DepositOptionMapper;
+import com.banklab.product.mapper.DepositProductMapper;
 import org.springframework.batch.core.StepContribution;
 import org.springframework.batch.core.scope.context.ChunkContext;
 import org.springframework.batch.core.step.tasklet.Tasklet;
@@ -10,18 +10,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
-public class DeleteCreditLoanTasklet implements Tasklet {
+public class DeleteDepositTasklet implements Tasklet {
 
     @Autowired
-    private CreditLoanProductMapper creditLoanProductMapper;
+    private DepositProductMapper depositProductMapper;
     
     @Autowired
-    private CreditLoanOptionMapper creditLoanOptionMapper;
+    private DepositOptionMapper depositOptionMapper;
 
     @Override
     public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
-        creditLoanProductMapper.deleteAllCreditLoanProducts();
-        creditLoanOptionMapper.deleteAllCreditLoanOptions();
+        depositProductMapper.deleteAllDepositProducts();
+        depositOptionMapper.deleteAllDepositOptions();
         return RepeatStatus.FINISHED;
     }
 }
