+++ src/main/java/com/banklab/product/domain/deposit/DepositProduct.java
@@ -1,4 +1,4 @@
-package com.banklab.product.domain;
+package com.banklab.product.domain.deposit;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
