+++ src/main/java/com/banklab/product/domain/savings/SavingsProduct.java
@@ -1,10 +1,11 @@
-package com.banklab.product.domain;
+package com.banklab.product.domain.savings;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
+import lombok.Builder;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -12,7 +13,7 @@ import java.time.LocalDateTime;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class CreditLoanProduct {
+public class SavingsProduct {
     private Long id;
     private String dclsMonth;
     private String finCoNo;
@@ -20,15 +21,14 @@ public class CreditLoanProduct {
     private String finPrdtCd;
     private String finPrdtNm;
     private String joinWay;
-    private String crdtPrdtType;
-    private String crdtPrdtTypeNm;
-    private String cbName;
+    private String mtrtInt;
+    private String spclCnd;
+    private Integer joinDeny;
+    private String joinMember;
+    private String etcNote;
+    private BigDecimal maxLimit;
     private LocalDate dclsStrtDay;
     private LocalDate dclsEndDay;
     private LocalDateTime finCoSubmDay;
-
-    // 위험도 분석용 추가 필드들
-    private String spclCnd; // 우대조건 (분석용)
-    private String etcNote; // 기타유의사항 (분석용)
-
 }
+
