+++ src/main/java/com/banklab/product/mapper/CreditLoanOptionMapper.java
@@ -1,6 +1,6 @@
 package com.banklab.product.mapper;
 
-import com.banklab.product.domain.CreditLoanOption;
+import com.banklab.product.domain.creditloan.CreditLoanOption;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
