+++ src/main/java/com/banklab/product/mapper/DepositOptionMapper.java
@@ -1,6 +1,6 @@
 package com.banklab.product.mapper;
 
-import com.banklab.product.domain.DepositOption;
+import com.banklab.product.domain.deposit.DepositOption;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
