+++ src/main/java/com/banklab/product/mapper/DepositProductMapper.java
@@ -1,6 +1,6 @@
 package com.banklab.product.mapper;
 
-import com.banklab.product.domain.DepositProduct;
+import com.banklab.product.domain.deposit.DepositProduct;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
