+++ src/main/java/com/banklab/product/service/ProductRateServiceImpl.java
@@ -5,6 +5,9 @@ import com.banklab.product.dto.ProductRateInfo;
 import com.banklab.product.dto.deposit.DepositWithOptionsDto;
 import com.banklab.product.dto.savings.SavingsWithOptionsDto;
 import com.banklab.product.dto.creditloan.CreditLoanWithOptionsDto;
+import com.banklab.product.service.creditloan.CreditLoanDetailService;
+import com.banklab.product.service.deposit.DepositDetailService;
+import com.banklab.product.service.savings.SavingsDetailService;
 import com.banklab.risk.domain.ProductRiskRating;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
