+++ src/main/java/com/banklab/product/service/creditloan/CreditLoanApiService.java
@@ -1,4 +1,4 @@
-package com.banklab.product.service;
+package com.banklab.product.service.creditloan;
 
 import com.banklab.product.dto.creditloan.CreditLoanApiResultWrapper;
 import com.banklab.product.dto.creditloan.CreditLoanProductAndOptionListDto;
