+++ src/main/java/com/banklab/product/service/deposit/DepositApiService.java
@@ -1,7 +1,7 @@
-package com.banklab.product.service;
+package com.banklab.product.service.deposit;
 
-import com.banklab.product.dto.savings.SavingsApiResultWrapper;
-import com.banklab.product.dto.savings.SavingsProductAndOptionListDto;
+import com.banklab.product.dto.deposit.DepositApiResultWrapper;
+import com.banklab.product.dto.deposit.DepositProductAndOptionListDto;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.*;
 import org.springframework.stereotype.Service;
@@ -12,20 +12,20 @@ import java.util.Collections;
 import java.util.stream.Collectors;
 
 /**
- * 적금 상품 fetch API
+ * 예금 상품 fetch API
  */
 @Service
-public class SavingsApiService {
+public class DepositApiService {
     private final RestTemplate restTemplate = new RestTemplate();
 
-    private final String API_URL = "https://finlife.fss.or.kr/finlifeapi/savingProductsSearch.json";
+    private final String API_URL = "https://finlife.fss.or.kr/finlifeapi/depositProductsSearch.json";
     @Value("${finlife.api-key}")
     private String API_KEY;
 
-    public SavingsProductAndOptionListDto fetchProductsFromApi() {
+    public DepositProductAndOptionListDto fetchProductsFromApi() {
         String fullUrl = API_URL + "?auth=" + API_KEY + "&topFinGrpNo=020000&pageNo=1";
 
-        System.out.println("Savings API 호출 URL: " + fullUrl);
+        System.out.println("Deposit API 호출 URL: " + fullUrl);
 
         try {
             // HTTP 헤더 설정
@@ -43,32 +43,32 @@ public class SavingsApiService {
                     String.class
             );
 
-            System.out.println("Savings API 응답 상태: " + stringResponse.getStatusCode());
-            System.out.println("Savings API 응답 Body: " + stringResponse.getBody());
+            System.out.println("Deposit API 응답 상태: " + stringResponse.getStatusCode());
+            System.out.println("Deposit API 응답 Body: " + stringResponse.getBody());
 
             if (stringResponse.getBody() == null || stringResponse.getBody().isEmpty()) {
-                throw new RuntimeException("Savings API 응답이 비어있습니다.");
+                throw new RuntimeException("Deposit API 응답이 비어있습니다.");
             }
 
-            // SavingsApiResultWrapper로 파싱
-            ResponseEntity<SavingsApiResultWrapper> response = restTemplate.exchange(
+            // DepositApiResultWrapper로 파싱
+            ResponseEntity<DepositApiResultWrapper> response = restTemplate.exchange(
                     fullUrl,
                     HttpMethod.GET,
                     entity,
-                    SavingsApiResultWrapper.class
+                    DepositApiResultWrapper.class
             );
 
             if (response.getBody() == null) {
-                throw new RuntimeException("Savings API 응답 파싱 실패. JSON 구조를 확인해주세요.");
+                throw new RuntimeException("Deposit API 응답 파싱 실패. JSON 구조를 확인해주세요.");
             }
 
-            SavingsApiResultWrapper.SavingsApiResult result = response.getBody().getResult();
+            DepositApiResultWrapper.DepositApiResult result = response.getBody().getResult();
 
             if (result == null) {
-                throw new RuntimeException("Savings API 결과 데이터가 null입니다.");
+                throw new RuntimeException("Deposit API 결과 데이터가 null입니다.");
             }
 
-            SavingsProductAndOptionListDto dto = new SavingsProductAndOptionListDto();
+            DepositProductAndOptionListDto dto = new DepositProductAndOptionListDto();
 
             dto.setProducts(result.getBaseList().stream()
                     .collect(Collectors.toList()));
@@ -76,17 +76,17 @@ public class SavingsApiService {
             dto.setOptions(result.getOptionList().stream()
                     .collect(Collectors.toList()));
 
-            System.out.println("savings API 호출 성공: 상품 " + result.getBaseList().size() + "개, 옵션 " + result.getOptionList().size() + "개");
+            System.out.println("Deposit API 호출 성공: 상품 " + result.getBaseList().size() + "개, 옵션 " + result.getOptionList().size() + "개");
 
             return dto;
 
         } catch (ResourceAccessException e) {
-            System.err.println("Savings API 네트워크 연결 오류: " + e.getMessage());
-            throw new RuntimeException("Savings API 서버에 연결할 수 없습니다: " + e.getMessage(), e);
+            System.err.println("Deposit API 네트워크 연결 오류: " + e.getMessage());
+            throw new RuntimeException("Deposit API 서버에 연결할 수 없습니다: " + e.getMessage(), e);
         } catch (Exception e) {
-            System.err.println("Savings API 호출 오류: " + e.getMessage());
+            System.err.println("Deposit API 호출 오류: " + e.getMessage());
             e.printStackTrace();
-            throw new RuntimeException("Savings API 호출 중 오류가 발생했습니다: " + e.getMessage(), e);
+            throw new RuntimeException("Deposit API 호출 중 오류가 발생했습니다: " + e.getMessage(), e);
         }
     }
 }
