+++ src/main/java/com/banklab/product/service/savings/SavingsDetailServiceImpl.java
@@ -1,7 +1,9 @@
-package com.banklab.product.service;
+package com.banklab.product.service.savings;
 
 import com.banklab.product.domain.*;
-import com.banklab.product.dto.deposit.*;
+import com.banklab.product.domain.savings.SavingsOption;
+import com.banklab.product.domain.savings.SavingsProduct;
+import com.banklab.product.dto.savings.*;
 import com.banklab.product.mapper.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -14,79 +16,83 @@ import java.util.stream.Collectors;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class DepositDetailServiceImpl implements DepositDetailService {
+public class SavingsDetailServiceImpl implements SavingsDetailService {
 
-    private final DepositProductMapper depositProductMapper;
-    private final DepositOptionMapper depositOptionMapper;
+    private final SavingsProductMapper savingsProductMapper;
+    private final SavingsOptionMapper savingsOptionMapper;
 
     @Override
-    public DepositWithOptionsDto getDepositWithOptions(String dclsMonth, String finCoNo, String finPrdtCd) {
+    public SavingsWithOptionsDto getSavingsWithOptions(String dclsMonth, String finCoNo, String finPrdtCd) {
         try {
-            log.info("예금 상품과 옵션 조회: dclsMonth={}, finCoNo={}, finPrdtCd={}", dclsMonth, finCoNo, finPrdtCd);
-
+            System.out.println("dcls"+dclsMonth+"finCoNo"+finCoNo+"finPrdtCd"+finPrdtCd);
+            log.info("적금 상품과 옵션 조회: dclsMonth={}, finCoNo={}, finPrdtCd={}", dclsMonth, finCoNo, finPrdtCd);
+            
             // 1. 상품 조회
-            DepositProduct product = depositProductMapper.findByProductKey(dclsMonth, finCoNo, finPrdtCd);
+            SavingsProduct product = savingsProductMapper.findByProductKey(dclsMonth, finCoNo, finPrdtCd);
             if (product == null) {
-                log.warn("예금 상품을 찾을 수 없음: dclsMonth={}, finCoNo={}, finPrdtCd={}", dclsMonth, finCoNo, finPrdtCd);
+                log.warn("적금 상품을 찾을 수 없음: dclsMonth={}, finCoNo={}, finPrdtCd={}", dclsMonth, finCoNo, finPrdtCd);
                 return null;
             }
 
             // 2. 옵션 조회
-            List<DepositOption> options = depositOptionMapper.findOptionsByProduct(dclsMonth, finCoNo, finPrdtCd);
-
+            List<SavingsOption> options = savingsOptionMapper.findOptionsByProduct(dclsMonth, finCoNo, finPrdtCd);
+            
             // 3. DTO 변환
-            List<DepositOptionDto> optionDTOs = options.stream()
-                    .map(this::convertToDepositOptionDTO)
+            List<SavingsOptionDto> optionDTOs = options.stream()
+                    .map(this::convertToSavingsOptionDto)
                     .collect(Collectors.toList());
-
+            
             // 4. 통합 DTO 빌드
-            return buildDepositWithOptionsDTO(product, optionDTOs);
-
+            return buildSavingsWithOptionsDto(product, optionDTOs);
+            
         } catch (Exception e) {
-            log.error("예금 상품 옵션 조회 중 오류 발생: dclsMonth={}, finCoNo={}, finPrdtCd={}",
-                    dclsMonth, finCoNo, finPrdtCd, e);
+            log.error("적금 상품 옵션 조회 중 오류 발생: dclsMonth={}, finCoNo={}, finPrdtCd={}", 
+                     dclsMonth, finCoNo, finPrdtCd, e);
             return null;
         }
     }
 
+
     /**
-     * DepositOption을 DepositOptionDTO로 변환
+     * SavingsOption을 SavingsOptionDTO로 변환
      */
-    private DepositOptionDto convertToDepositOptionDTO(DepositOption option) {
-        return DepositOptionDto.builder()
+    private SavingsOptionDto convertToSavingsOptionDto(SavingsOption option) {
+        return SavingsOptionDto.builder()
                 .id(option.getId())
                 .intrRateType(option.getIntrRateType())
                 .intrRateTypeNm(option.getIntrRateTypeNm())
                 .saveTrm(String.valueOf(option.getSaveTrm()))
                 .intrRate(option.getIntrRate())
                 .intrRate2(option.getIntrRate2())
+                .rsrvType(option.getRsrvType())
+                .rsrvTypeNm(option.getRsrvTypeNm())
                 .build();
     }
 
     /**
-     * DepositProduct와 옵션들로 DepositWithOptionsDTO 빌드
+     * SavingsProduct와 옵션들로 SavingsWithOptionsDTO 빌드
      */
-    private DepositWithOptionsDto buildDepositWithOptionsDTO(DepositProduct product, List<DepositOptionDto> options) {
+    private SavingsWithOptionsDto buildSavingsWithOptionsDto(SavingsProduct product, List<SavingsOptionDto> options) {
         // 최저/최고 금리 계산
         BigDecimal minRate = options.stream()
-                .map(DepositOptionDto::getIntrRate)
+                .map(SavingsOptionDto::getIntrRate)
                 .filter(rate -> rate != null)
                 .min(BigDecimal::compareTo)
                 .orElse(BigDecimal.ZERO);
         
         BigDecimal maxRate = options.stream()
-                .map(DepositOptionDto::getIntrRate2)
+                .map(SavingsOptionDto::getIntrRate2)
                 .filter(rate -> rate != null)
                 .max(BigDecimal::compareTo)
                 .orElse(BigDecimal.ZERO);
 
-        return DepositWithOptionsDto.builder()
+        return SavingsWithOptionsDto.builder()
                 .dclsMonth(product.getDclsMonth())
                 .finCoNo(product.getFinCoNo())
                 .finPrdtCd(product.getFinPrdtCd())
                 .finPrdtNm(product.getFinPrdtNm())
                 .korCoNm(product.getKorCoNm())
-                .productType(ProductType.DEPOSIT)
+                .productType(ProductType.SAVINGS)
                 .joinWay(product.getJoinWay())
                 .mtrtInt(product.getMtrtInt())
                 .spclCnd(product.getSpclCnd())
