+++ src/main/java/com/banklab/typetest/service/ProductRecommendationServiceImpl.java
@@ -5,9 +5,9 @@ import com.banklab.product.service.ProductRateService;
 import com.banklab.product.mapper.DepositProductMapper;
 import com.banklab.product.mapper.SavingsProductMapper;
 import com.banklab.product.mapper.CreditLoanProductMapper;
-import com.banklab.product.domain.DepositProduct;
-import com.banklab.product.domain.SavingsProduct;
-import com.banklab.product.domain.CreditLoanProduct;
+import com.banklab.product.domain.deposit.DepositProduct;
+import com.banklab.product.domain.savings.SavingsProduct;
+import com.banklab.product.domain.creditloan.CreditLoanProduct;
 import com.banklab.risk.domain.ProductRiskRating;
 import com.banklab.risk.domain.RiskLevel;
 import com.banklab.risk.mapper.ProductRiskRatingMapper;
