+++ src/main/resources/com/banklab/product/mapper/DepositProductMapper.xml
@@ -4,7 +4,7 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="com.banklab.product.mapper.DepositProductMapper">
-    <insert id="insertDepositProduct" parameterType="com.banklab.product.domain.DepositProduct">
+    <insert id="insertDepositProduct" parameterType="com.banklab.product.domain.deposit.DepositProduct">
         INSERT INTO deposit_product (
             dcls_month, fin_co_no, kor_co_nm, fin_prdt_cd, fin_prdt_nm, join_way, mtrt_int, spcl_cnd, join_deny,
             join_member, etc_note, max_limit, dcls_strt_day, dcls_end_day, fin_co_subm_day
@@ -13,13 +13,13 @@
             #{joinMember}, #{etcNote}, #{maxLimit}, #{dclsStrtDay}, #{dclsEndDay}, #{finCoSubmDay}
         )
     </insert>
-    <select id="findAllDepositProducts" resultType="com.banklab.product.domain.DepositProduct">
+    <select id="findAllDepositProducts" resultType="com.banklab.product.domain.deposit.DepositProduct">
         SELECT * FROM deposit_product
     </select>
     <delete id="deleteAllDepositProducts">
         DELETE FROM deposit_product
     </delete>
-    <select id="findByProductKey" resultType="com.banklab.product.domain.DepositProduct">
+    <select id="findByProductKey" resultType="com.banklab.product.domain.deposit.DepositProduct">
         SELECT 
             id,
             dcls_month as dclsMonth,
@@ -43,7 +43,7 @@
           AND fin_prdt_cd = #{finPrdtCd}
     </select>
 
-    <select id="findById" resultType="com.banklab.product.domain.DepositProduct">
+    <select id="findById" resultType="com.banklab.product.domain.deposit.DepositProduct">
         SELECT 
             id,
             dcls_month as dclsMonth,
