+++ src/main/resources/com/banklab/product/mapper/SavingsOptionMapper.xml
@@ -5,7 +5,7 @@
 
 <mapper namespace="com.banklab.product.mapper.SavingsOptionMapper">
     
-    <resultMap id="SavingsOptionResultMap" type="com.banklab.product.domain.SavingsOption">
+    <resultMap id="SavingsOptionResultMap" type="com.banklab.product.domain.savings.SavingsOption">
         <id property="id" column="id"/>
         <result property="dclsMonth" column="dcls_month"/>
         <result property="finCoNo" column="fin_co_no"/>
@@ -19,7 +19,7 @@
         <result property="intrRate2" column="intr_rate2"/>
     </resultMap>
 
-    <insert id="insertSavingsOption" parameterType="com.banklab.product.domain.SavingsOption">
+    <insert id="insertSavingsOption" parameterType="com.banklab.product.domain.savings.SavingsOption">
         INSERT INTO savings_option (
             dcls_month, fin_co_no, fin_prdt_cd,
             intr_rate_type, intr_rate_type_nm, rsrv_type, rsrv_type_nm,
@@ -39,7 +39,7 @@
         FROM savings_option
         ORDER BY fin_co_no, fin_prdt_cd, save_trm
     </select>
-    <select id="findOptionsByProduct" resultType="com.banklab.product.domain.SavingsOption">
+    <select id="findOptionsByProduct" resultType="com.banklab.product.domain.savings.SavingsOption">
         SELECT
             id,
             dcls_month as dclsMonth,
